separate shell from variableexplorer
