+++ spyder/plugins/ipythonconsole/widgets/client.py
@@ -476,10 +476,6 @@ class ClientWidget(QWidget, SaveHistoryMixin, SpyderWidgetMixin):
         # To update history after execution
         self.shellwidget.executed.connect(self.update_history)
 
-        # To update the Variable Explorer after execution
-        self.shellwidget.executed.connect(
-            self.shellwidget.refresh_namespacebrowser)
-
         # To enable the stop button when executing a process
         self.shellwidget.executing.connect(
             self.sig_execution_state_changed)
