+++ spyder/plugins/variableexplorer/plugin.py
@@ -68,5 +68,8 @@ class VariableExplorer(SpyderDockablePlugin, ShellConnectMixin):
     # ------------------------------------------------------------------------
     def on_connection_to_external_spyder_kernel(self, shellwidget):
         """Send namespace view settings to the kernel."""
-        shellwidget.set_namespace_view_settings()
-        shellwidget.refresh_namespacebrowser()
+        widget = self.get_widget_for_shellwidget(shellwidget)
+        if widget is None:
+            return
+        widget.set_namespace_view_settings()
+        widget.refresh_namespacebrowser()
