+++ spyder/plugins/variableexplorer/widgets/main_widget.py
@@ -426,6 +426,13 @@ class VariableExplorerWidget(ShellConnectMainWidget):
         nsb.set_shellwidget(shellwidget)
         nsb.setup()
         self._set_actions_and_menus(nsb)
+
+        # To update the Variable Explorer after execution
+        shellwidget.executed.connect(
+            nsb.refresh_namespacebrowser)
+        shellwidget.sig_kernel_has_started.connect(
+            nsb.on_kernel_started)
+
         return nsb
 
     def close_widget(self, nsb):
@@ -438,6 +445,10 @@ class VariableExplorerWidget(ShellConnectMainWidget):
             self.start_spinner)
         nsb.sig_stop_spinner_requested.disconnect(
             self.stop_spinner)
+        nsb.shellwidget.executed.disconnect(
+            nsb.refresh_namespacebrowser)
+        nsb.shellwidget.sig_kernel_has_started.disconnect(
+            nsb.on_kernel_started)
         nsb.close()
 
     def import_data(self, filenames=None):
